git-svn-id: https://svn.macports.org/repository/macports/branches/gsoc15-portfile/base@139406 d073be05-634f-4543-b044-5fe20cf6d1d6

+++ src/port1.0/portuninstall.tcl
@@ -49,8 +49,8 @@ options uninstall.asroot
 default uninstall.asroot no
 
 proc portuninstall::uninstall_start {args} {
-    global prefix
-    if {![file writable $prefix] || ([getuid] == 0 && [geteuid] != 0)} {
+    global prefix ports_dryrun
+    if {(![file writable $prefix] && ![tbool ports_dryrun]) || ([getuid] == 0 && [geteuid] != 0)} {
         # if install location is not writable, need root privileges
         elevateToRoot "uninstall"
     }
